rationHP

+++ src2D/datahelpers/dataloaders/fusedDataLoader.py
@@ -16,7 +16,7 @@ class FusionDataset(data.Dataset):
 		self.dataset_posetrack = posetrack(split, opts)
 
 		self.nVideos_posetrack = len(self.dataset_posetrack)
-		self.nVideos_h36m = min(len(self.dataset_h36m), (self.nVideos_posetrack)*opts.ratioHP)
+		self.nVideos_h36m = min(len(self.dataset_h36m), int((self.nVideos_posetrack)*opts.ratioHP))
 
 		print("Built h36m and posetrack dataset containing %d and %d samples" %(self.nVideos_h36m, self.nVideos_posetrack))
